+++ Sources/OpenFCPXMLKit/Extraction/Context/FCPXMLFrameRateSource.swift
@@ -10,7 +10,6 @@
 
 import Foundation
 import SwiftTimecode
-import SwiftExtensions
 
 extension FinalCutPro.FCPXML {
     /// Frame rate source for an extracted element.