+++ Sources/OpenFCPXMLKit/Model/Clips/FCPXMLLiveDrawing.swift
@@ -10,7 +10,6 @@
 
 import Foundation
 import SwiftTimecode
-import SwiftExtensions
 
 extension FinalCutPro.FCPXML {
     /// Live-drawing clip (FCPXML 1.11+).