+++ Sources/OpenFCPXMLKit/Model/Protocols/FCPXMLElementAudioRoleSourceChildren.swift
@@ -9,7 +9,6 @@
 //
 
 import Foundation
-import SwiftExtensions
 import SwiftTimecode
 
 public protocol FCPXMLElementAudioRoleSourceChildren: FCPXMLElement {