+++ Sources/OpenFCPXMLKit/Model/Resources/FCPXMLAsset.swift
@@ -10,7 +10,6 @@
 
 import Foundation
 import SwiftTimecode
-import SwiftExtensions
 
 extension FinalCutPro.FCPXML {
     /// Asset shared resource.