+++ Sources/OpenFCPXMLKit/Model/Resources/FCPXMLMediaMulticamAngle.swift
@@ -10,7 +10,6 @@
 
 import Foundation
 import SwiftTimecode
-import SwiftExtensions
 
 extension FinalCutPro.FCPXML.Media.Multicam {
     /// A container of story elements organized sequentially in time.