+++ Sources/OpenFCPXMLKit/Model/Roles/FCPXMLCaptionRole.swift
@@ -9,7 +9,7 @@
 //
 
 import Foundation
-import SwiftExtensions
+internal import SwiftExtensions
 
 extension FinalCutPro.FCPXML {
     /// Caption role.